% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOPS.R
\name{SOPS}
\alias{SOPS}
\title{The SOPS procedure}
\usage{
SOPS(X, Y, h_seq = NULL, n_ss = 20, kappa = 1, q = 0.2, thres.1 = TRUE)
}
\arguments{
\item{X}{design points, should be standardized in (0, 1)}

\item{Y}{responses}

\item{h_seq}{candidates for the bandwidth}

\item{n_ss}{times of sample-splitting for bandwidth selection}

\item{kappa}{controlling the size of removed neighborhood of detected jumps}

\item{q}{nominal FDR level}

\item{thres.1}{whether to increment the number of negatives by 1 to determine the threshold}
}
\value{
a list contains the number, locations and sizes of detected jumps
}
\description{
The SOPS procedure
}
\examples{
NA
}
